RUBY-1492 Add before_suite to force loading and instrumentation of Puma::Rack
This is pretty ugly, but I'm not sure there are any other good ways
around this.  Requiring puma in your Gemfile will not cause Puma::Rack
to be loaded, you actually need to using Puma to serve your application
for this to happen.  Puma::Rack is required by puma/configuration, which
gets included in either puma/cli or puma/control_cli.  In situations where
you are actually using Puma to serve your app Puma::Rack will have been loaded
prior to the agent and DependecyDetection will install our instrumentation.
For our multiverse suite  we need to make up for this situation, and the
before_suite option is the least terrible one that I know of at this time.