RUBY-1494 Ensure Puma::Rack instrumentation works without Rack
Most Ruby frameworks have a dependency on Rack so it is unlikely that many
people will run an application where only Puma::Rack is present, but it is
technically possible, so our instrumentation should handle that case. To
that end we needed an environment in our multiverse suite that ensured that
we properly instrument middleware without rack available.  Additionally,
in the MiddlewareTracing module we must conditionally pass a Rack::Request
to Transaction#start if Rack is present. Many tests in multiverse
needed to have their conditional statements moved above the require statements
as some of the files required reference Rack.